refactor: move metrics funcs to own file
