+++ cmd/kar-controllers/app/server.go
@@ -26,12 +26,10 @@ import (
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 
+	"github.com/project-codeflare/multi-cluster-app-dispatcher/cmd/kar-controllers/app/metrics"
 	"github.com/project-codeflare/multi-cluster-app-dispatcher/cmd/kar-controllers/app/options"
 	"github.com/project-codeflare/multi-cluster-app-dispatcher/pkg/controller/queuejob"
 	"github.com/project-codeflare/multi-cluster-app-dispatcher/pkg/health"
-	"github.com/prometheus/client_golang/prometheus"
-	"github.com/prometheus/client_golang/prometheus/collectors"
-	"github.com/prometheus/client_golang/prometheus/promhttp"
 
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 
@@ -40,8 +38,6 @@ import (
 	"github.com/project-codeflare/multi-cluster-app-dispatcher/pkg/config"
 )
 
-// Global Prometheus Registry
-var globalPromRegistry = prometheus.NewRegistry()
 
 func buildConfig(master, kubeconfig string) (*rest.Config, error) {
 	if master != "" || kubeconfig != "" {
@@ -87,20 +83,6 @@ func Run(ctx context.Context, opt *options.ServerOption) error {
 	return nil
 }
 
-// metricsHandler returns a http.Handler that serves the prometheus metrics
-func prometheusHandler() http.Handler {
-	// Add Go module build info.
-	globalPromRegistry.MustRegister(collectors.NewBuildInfoCollector())
-	globalPromRegistry.MustRegister(collectors.NewGoCollector())
-	globalPromRegistry.MustRegister(collectors.NewProcessCollector(collectors.ProcessCollectorOpts{}))
-
-	handlerOpts := promhttp.HandlerOpts{
-		ErrorHandling: promhttp.HTTPErrorOnError,
-	}
-
-	return promhttp.HandlerFor(globalPromRegistry, handlerOpts)
-}
-
 func healthHandler() http.Handler {
 	healthHandler := http.NewServeMux()
 	healthHandler.Handle("/healthz", &health.Handler{})
@@ -112,7 +94,7 @@ func startHealthAndMetricsServers(ctx context.Context, opt *options.ServerOption
 	g, ctx := errgroup.WithContext(ctx)
 
 	// metrics server
-	metricsServer, err := NewServer(opt.MetricsListenPort, "/metrics", prometheusHandler())
+	metricsServer, err := NewServer(opt.MetricsListenPort, "/metrics", metrics.PrometheusHandler())
 	if err != nil {
 		return err
 	}
